Moving over more of the server code to use typescript
