+++ package.json
@@ -8,7 +8,7 @@
     "postinstall": "typings install && npm run build",
     "prodBuild": "npm run build && node server/app.js",
     "runTypescriptBuild": "tsc && node dist/app",
-    "start": "concurrently \"webpack --env=development --watch --colors\" \"npm run runTypescriptBuild && node dist/app\""
+    "start": "concurrently \"webpack --env=development --watch --colors\" \"npm run runTypescriptBuild\""
   },
   "author": "Juan Cortez",
   "repository": {
